autoGenerateMatrixAndNewGraphVis

+++ src/main/java/ru/itmo/sortvis/GraphModelListener.java
@@ -1,8 +1,6 @@
 package ru.itmo.sortvis;
 
-import java.awt.*;
-
 public interface GraphModelListener {
-    void someChanges(byte[][] matrix, int[] coord);
+    void someChanges(byte[][] matrix, double[] coord);
 
 }
